Remove edge
