--- backend/app/serializers/api/edge/tmdb_movie_serializer.rb
@@ -1,13 +1,0 @@
-module Api
-  module Edge
-    class TmdbMovieSerializer < ApplicationSerializer
-      type 'tmdbMovies'
-
-      attributes :id,
-        :title,
-        :overview,
-        :release_date,
-        :poster_path
-    end
-  end
-end
