--- backend/app/serializers/api/edge/user_auth_serializer.rb
@@ -1,10 +1,0 @@
-module Api
-  module Edge
-    class UserAuthSerializer < ApplicationSerializer
-      type 'users'
-
-      attributes :username,
-        :email
-    end
-  end
-end
