+++ backend/app/serializers/api/v1/tmdb_movie_serializer.rb
@@ -1,5 +1,13 @@
 module Api
   module V1
-    class TmdbMovieSerializer < Edge::TmdbMovieSerializer; end
+    class TmdbMovieSerializer < ApplicationSerializer
+      type 'tmdbMovies'
+
+      attributes :id,
+        :title,
+        :overview,
+        :release_date,
+        :poster_path
+    end
   end
 end
