+++ backend/app/serializers/api/v1/tmdb_person_serializer.rb
@@ -1,5 +1,15 @@
 module Api
   module V1
-    class TmdbPersonSerializer < Edge::TmdbPersonSerializer; end
+    class TmdbPersonSerializer < ApplicationSerializer
+      type 'tmdbPeople'
+
+      attributes :id,
+        :name,
+        :known_for_department,
+        :profile_path
+
+      has_many :known_for,
+        serializer: versioned_class(TmdbMovieSerializer)
+    end
   end
 end
