+++ backend/app/serializers/api/v1/user_auth_serializer.rb
@@ -1,5 +1,10 @@
 module Api
   module V1
-    class UserAuthSerializer < Edge::UserAuthSerializer; end
+    class UserAuthSerializer < ApplicationSerializer
+      type 'users'
+
+      attributes :username,
+        :email
+    end
   end
 end
